requirements.txt
beautifulsoup4==4.14.3
chromadb==1.3.6
langchain==1.1.3
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.1.3
langchain_huggingface==1.1.0
langchain_mistralai==1.1.1
langchain_openai==1.1.2
langchain_text_splitters==1.0.0
langgraph==1.0.4
langsmith==0.4.56
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.52.1
tavily_python==0.7.17
